Problem: memiavl snapshot format is not optimal (#890)
* Problem: (memiavl) offset table is not used to reduce node size

Solution:
- store offset table inside keys/values file, so we can reference it with leaf index in nodes.
- leverage the property of post-order traversal to further reduce some offsets.
- specialize `Get` operation for persisted node to improve query performance.
- add native-endian implementation to improve performance.

in the end, the on-disk IAVL tree query performance is on-par with tidwall/btree library
which is used for cache store.

* cleanup and fix lint

* Apply suggestions from code review


* use simple delta encoding for key offsets

* new tricks to get back performance

* cleanup

* build recsplit index

* fix lint

* cleanup

* Apply suggestions from code review


* go mod tidy

---------

+++ versiondb/client/restore_app_db.go
@@ -283,7 +283,7 @@ func encodeNode(w io.Writer, node memiavl.PersistedNode) error {
 	if _, err := w.Write(buf[:n]); err != nil {
 		return err
 	}
-	n = binary.PutVarint(buf[:], node.Version())
+	n = binary.PutVarint(buf[:], int64(node.Version()))
 	if _, err := w.Write(buf[:n]); err != nil {
 		return err
 	}
